merge: resolve conflicts with upstream recordings refactor
Merge upstream refactoring (bbd5e09) that split recording writers into
platform-specific implementations. Resolved conflict in
platform-recording-writer.ts by adopting the abstract `normalizeChunk()`
method pattern introduced by the upstream, which replaces the inline
conditional normalization approach from the original PR commits.

Agent-Logs-Url: https://github.com/UseStitch/stitch/sessions/893e8fa7-108d-4e9a-b2a5-3bad1b542678
