+++ apps/web/src/components/navigation/activity-bar.tsx
@@ -4,6 +4,7 @@ import { Link, useRouterState } from '@tanstack/react-router';
 
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
 import { useDialogContext } from '@/context/dialog-context';
+import { useFullScreen } from '@/hooks/ui/use-fullscreen';
 import { cn } from '@/lib/utils';
 
 type ActivityItem = {
@@ -60,9 +61,20 @@ export function ActivityBar() {
   const routerState = useRouterState();
   const currentPath = routerState.location.pathname;
   const { setSettingsTab } = useDialogContext();
+  const isMac = window.electron?.platform === 'darwin';
+  const isFullScreen = useFullScreen();
+  const showTrafficLightPadding = isMac && !isFullScreen;
 
   return (
-    <div className="flex w-14 flex-col items-center border-r border-border/50 bg-sidebar px-1.5 pt-3 pb-3">
+    <div
+      className={cn(
+        'relative flex w-14 flex-col items-center bg-sidebar px-1.5 pb-3',
+        showTrafficLightPadding ? 'pt-10' : 'border-r border-border/50 pt-3',
+      )}
+    >
+      {showTrafficLightPadding && (
+        <div className="pointer-events-none absolute top-9 right-0 bottom-0 border-r border-border/50" />
+      )}
       <div className="flex w-full flex-col items-center gap-2">
         {ACTIVITY_ITEMS.map((item) => {
           const active = isActive(item.matchPrefix, currentPath);
