+++ packages/shared/package.json
@@ -32,6 +32,7 @@
     "@ai-sdk/provider": "catalog:ai",
     "ai": "catalog:ai",
     "typescript": "catalog:",
-    "vitest": "catalog:testing"
+    "vitest": "catalog:testing",
+    "zod": "catalog:"
   }
 }
